Se agrego el lanzamiento de excepciones

+++ src/main/java/com/libreria/libreria/service/LibroService.java
@@ -25,7 +25,16 @@ public class LibroService {
     private EditorialRepository editorialRepository;
 
     @Transactional
-    public void create(String titulo, Integer anio, Integer ejemplares, Long isbn, String autor, String editorial) {
+    public void create(String titulo, Integer anio, Integer ejemplares, Long isbn, String autor, String editorial) throws Exception {
+        
+        if(!editorialRepository.existsById(editorial)){
+            throw new Exception("La editorial no existe");
+        }
+
+        if(!autorRepository.existsById(autor)){
+            throw new Exception("El autor no existe");
+        }
+        
         Libro libro = new Libro();
         
         libro.setAlta(true);
@@ -47,12 +56,58 @@ public class LibroService {
     }
 
     @Transactional
-    public void delete(String id) {
+    public Libro getLibro(String id) throws Exception {
+        if(!libroRepository.existsById(id)){
+            throw new Exception("El libro no existe");
+        }
+        return libroRepository.getById(id);
+    }
+
+    @Transactional
+    public void modificar (String id, String titulo, Integer anio, Integer ejemplares, Long isbn, String autor, String editorial) throws Exception {
+
+        if(!libroRepository.existsById(id)){
+            throw new Exception("El libro no existe");
+        }
+
+        if(!editorialRepository.existsById(editorial)){
+            throw new Exception("La editorial no existe");
+        }
+
+        if(!autorRepository.existsById(autor)){
+            throw new Exception("El autor no existe");
+        }
+
+        Libro libro = libroRepository.getById(id);
+
+        libro.setTitulo(titulo);
+        libro.setAnio(anio);
+        libro.setEjemplares(ejemplares);
+        libro.setIsbn(isbn);
+        libro.setAutor(autorRepository.findByIdAutor(autor));
+        libro.setEditorial(editorialRepository.findByIdEditorial(editorial));
+        
+        libroRepository.save(libro);
+
+    }
+
+    @Transactional
+    public void delete(String id) throws Exception {
+
+        if(!libroRepository.existsById(id)){
+            throw new Exception("El libro no existe");
+        }
+
         libroRepository.deleteById(id);
     }
 
     @Transactional
-    public void alta(String id) {
+    public void alta(String id) throws Exception {
+
+        if(!libroRepository.existsById(id)){
+            throw new Exception("El libro no existe");
+        }
+
         Libro libro = libroRepository.getById(id);
         libro.setAlta(true);
         libroRepository.save(libro);
